implemented 50% main features
Description: Compiler basics added. vite.config.js changed to access .env. Code Editor and Submit Button read from an added useState and passes it to compiler-services. Added redirect for user, and empty SearchHut page. Todo: cleaning up the compile features so that things display on the side, and adding the ability to save a script to your personal hut.